% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_samples.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract samples from a oncoSimulIndiv object}
\usage{
extract_samples(x, opt)
}
\arguments{
\item{x}{An oncoSimulIndiv object}

\item{opt}{A list with the options of the simulations
The needed elements are:
det.limit
nNS_pos
nNS_neg
nNS_neu
nS
sNS_pos
sNS_neg}
}
\value{
A tibble with different papameters for all the sampled times
}
\description{
This function was adapted from OncoSimulR::get.mut.vec
}
