% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pNpS.sims.R
\docType{package}
\name{pNpS.sims}
\alias{pNpS.sims}
\alias{pNpS.sims-package}
\title{pNpS.sims A package for the simulation of the evolution of pN/pS on tumors}
\description{
This package provides all the code needed to reproduce the results
}
\section{pNPS.sims functions}{

}

