% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_samples.R
\name{extract_samples}
\alias{extract_samples}
\title{Extract samples from a oncoSimulIndiv object}
\usage{
extract_samples(x, det.limit = 0.05, nNS, nS, finalTime, keepEvery)
}
\arguments{
\item{x}{An oncoSimulIndiv object}

\item{det.limit}{A numeric value for the frquency threshold of a detected mutation}

\item{nNS}{Number of NS (Non-Synonymous) loci}

\item{nS}{Number of NS (Synonymous) loci}
}
\value{
A tibble with different papameters for all the sampled times
}
\description{
Extract samples from a oncoSimulIndiv object
}
